+++ idea-plugin/src/com/reason/lang/reason/RmlParser.java
@@ -574,6 +574,15 @@ public class RmlParser extends CommonParser<RmlTypes> {
                     advance().
                     remapCurrentToken(m_types.TAG_NAME).
                     wrapWith(nextTokenType == m_types.UIDENT ? m_types.C_UPPER_SYMBOL : m_types.C_LOWER_SYMBOL);
+        } else if (nextTokenType == m_types.GT) {
+            // a React fragment start
+            state.remapCurrentToken(m_types.TAG_LT).
+                    mark(jsxTag, m_types.C_TAG).
+                    mark(jsxStartTag, m_types.C_TAG_START).
+                    advance().
+                    remapCurrentToken(m_types.TAG_GT).
+                    advance().
+                    popEnd();
         }
     }
 
@@ -586,11 +595,19 @@ public class RmlParser extends CommonParser<RmlTypes> {
                 state.popEnd();
             }
 
-            state.remapCurrentToken(m_types.TAG_LT).
+            state.remapCurrentToken(m_types.TAG_LT_SLASH).
                     mark(jsxTagClose, m_types.C_TAG_CLOSE).
                     advance().
                     remapCurrentToken(m_types.TAG_NAME).
                     wrapWith(nextTokenType == m_types.UIDENT ? m_types.C_UPPER_SYMBOL : m_types.C_LOWER_SYMBOL);
+        } else if (nextTokenType == m_types.GT) {
+            // a React fragment end
+            state.remapCurrentToken(m_types.TAG_LT_SLASH).
+                    mark(jsxTagClose, m_types.C_TAG_CLOSE).
+                    advance().
+                    remapCurrentToken(m_types.TAG_GT).
+                    advance().
+                    popEnd();
         }
     }